requirements.txt
streamlit
openai
semantic-router
python-dotenv
